# uzum_api_gateway/api/docs/swagger.yaml
definitions:
...
  catalog_service.CreateProduct:
    properties:
      active:
        type: boolean
      description_en:
        type: string
      description_ru:
        type: string
      description_uz:
        type: string
      discount_percent:
        type: number
      image:
        items:
          type: string
        type: array
      in_price:
        type: number
      left_count:
        type: integer
      name_en:
        type: string
      name_ru:
        type: string
      name_uz:
        type: string
      order_no:
        type: integer
      out_price:
        type: number
    type: object
...
  catalog_service.Empty:
    type: object
...
  catalog_service.GetAllProductResponse:
    properties:
      Count:
        type: integer
      Products:
        items:
          $ref: '#/definitions/catalog_service.Product'
        type: array
    type: object
...
  catalog_service.Product:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      deleted_at:
        type: integer
      description_en:
        type: string
      description_ru:
        type: string
      description_uz:
        type: string
      discount_percent:
        type: number
      id:
        type: string
      image:
        items:
          type: string
        type: array
      in_price:
        type: number
      left_count:
        type: integer
      name_en:
        type: string
      name_ru:
        type: string
      name_uz:
        type: string
      order_no:
        type: integer
      out_price:
        type: number
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  catalog_service.UpdateProduct:
    properties:
      active:
        type: boolean
      description_en:
        type: string
      description_ru:
        type: string
      description_uz:
        type: string
      discount_percent:
        type: number
      id:
        type: string
      image:
        items:
          type: string
        type: array
      in_price:
        type: number
      left_count:
        type: integer
      name_en:
        type: string
      name_ru:
        type: string
      name_uz:
        type: string
      order_no:
        type: integer
      out_price:
        type: number
    type: object
...
  models.ResponseError:
    properties:
      error: {}
    type: object
...
  user_service.CreateCustomer:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      firstname:
        type: string
      gender:
        type: string
      languange:
        type: string
      lastname:
        type: string
      phone:
        type: string
    type: object
...
  user_service.Customer:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      deleted_at:
        type: integer
      email:
        type: string
      firstname:
        type: string
      gender:
        type: string
      id:
        type: string
      languange:
        type: string
      lastname:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  user_service.Empty:
    type: object
...
  user_service.GetListCustomerResponse:
    properties:
      Count:
        type: integer
      Customers:
        items:
          $ref: '#/definitions/user_service.Customer'
        type: array
    type: object
...
  user_service.UpdateCustomer:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      languange:
        type: string
      lastname:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /createCustomer:
    post:
      consumes:
      - application/json
      description: API for creating customer
      parameters:
      - description: customer
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateCustomer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.CreateCustomer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create customer
      tags:
      - customer
...
  /deleteCustomer/{id}:
    delete:
      consumes:
      - application/json
      description: API for deleting a customer by ID
      parameters:
      - description: customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Empty'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a customer by ID
      tags:
      - customer
...
  /getallproduct:
    get:
      consumes:
      - application/json
      description: API for getting all products
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.GetAllProductResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get All Products
      tags:
      - product
...
  /getbyidcustomer/{id}:
    get:
      consumes:
      - application/json
      description: API for getting a single category by ID
      parameters:
      - description: customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Customer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a single customer by ID
      tags:
      - customer
...
  /getlistcustomer:
    get:
      consumes:
      - application/json
      description: API for getting list customer
      parameters:
      - description: customers
        in: query
        name: customer
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.GetListCustomerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get List Customer
      tags:
      - customer
  /product:
    post:
      consumes:
      - application/json
      description: API for creating a product
      parameters:
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/catalog_service.CreateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.Product'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a product
      tags:
      - product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: API for deleting a product by ID
      parameters:
      - description: product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.Empty'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a product by ID
      tags:
      - product
    get:
      consumes:
      - application/json
      description: API for getting a single product by ID
      parameters:
      - description: product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.Product'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a single product by ID
      tags:
      - product
    put:
      consumes:
      - application/json
      description: API for updating a product by ID
      parameters:
      - description: product ID
        in: path
        name: id
        required: true
        type: string
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/catalog_service.UpdateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.Product'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update a product by ID
      tags:
      - product
...
  /updateCustomer/{id}:
    put:
      consumes:
      - application/json
      description: API for updating a customer by ID
      parameters:
      - description: customer ID
        in: path
        name: id
        required: true
        type: string
      - description: customer
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateCustomer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Customer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update a customer by ID
      tags:
      - customer
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"